+++ packages/cli-utils/src/index.ts
@@ -6,16 +6,12 @@ import { printSchema } from 'graphql';
 
 import type { GraphQLSchema } from 'graphql';
 import type { TsConfigJson } from 'type-fest';
-import {
-  resolveTypeScriptRootDir,
-  load,
-  check,
-  type FormattedDisplayableDiagnostic,
-} from '@gql.tada/internal';
+import { load } from '@gql.tada/internal';
 
-import type { GraphQLSPConfig } from './lsp';
 import { getGraphQLSPConfig } from './lsp';
 import { ensureTadaIntrospection } from './tada';
+import { getTsConfig } from './tsconfig';
+import { check } from './commands/check';
 
 interface GenerateSchemaOptions {
   headers?: Record<string, string>;
@@ -78,7 +74,12 @@ export async function generateSchema(
 }
 
 export async function generateTadaTypes(shouldPreprocess = false, cwd: string = process.cwd()) {
-  const config = await getConfig();
+  const tsConfig = await getTsConfig();
+  if (!tsConfig) {
+    return;
+  }
+
+  const config = getGraphQLSPConfig(tsConfig);
   if (!config) {
     return;
   }
@@ -140,93 +141,15 @@ async function main() {
       return generateTadaTypes(shouldPreprocess);
     })
     .command('check')
-    .option('--severity', 'The minimum severity of diagnostics to display.')
+    .option('--level', 'The minimum severity of diagnostics to display (error | warn | info).')
+    .option('--exit-on-warn', 'Whether to exit with a non-zero code when there are warnings.')
     .action(async (opts) => {
-      const config = await getConfig();
-      if (!config) {
-        return;
-      }
-      const result = (await check(config, opts.severity || 'error')) || [];
-      const errorDiagnostics = result.filter((d) => d.severity === 'error');
-      const warnDiagnostics = result.filter((d) => d.severity === 'warn');
-      const infoDiagnostics = result.filter((d) => d.severity === 'info');
-      if (
-        errorDiagnostics.length === 0 &&
-        warnDiagnostics.length === 0 &&
-        infoDiagnostics.length === 0
-      ) {
-        // eslint-disable-next-line no-console
-        console.log('No issues found! 🎉');
-        return;
-      } else {
-        const errorReport = errorDiagnostics.length
-          ? `Found ${errorDiagnostics.length} Errors:\n\n${constructDiagnosticsPerFile(
-              errorDiagnostics
-            )}\n\n`
-          : ``;
-        const warningsReport = warnDiagnostics.length
-          ? `Found ${warnDiagnostics.length} Warnings:\n\n${constructDiagnosticsPerFile(
-              warnDiagnostics
-            )}\n\n`
-          : ``;
-        const suggestionsReport = infoDiagnostics.length
-          ? `Found ${infoDiagnostics.length} Suggestions:\n\n${constructDiagnosticsPerFile(
-              infoDiagnostics
-            )}\n\n`
-          : ``;
-        // eslint-disable-next-line no-console
-        console.log(`${errorReport}${warningsReport}${suggestionsReport}`);
-      }
+      check({
+        exitOnWarn: opts['exit-on-warn'] !== undefined ? opts['exit-on-warn'] : false,
+        minSeverity: opts.level || 'error',
+      });
     });
   prog.parse(process.argv);
 }
-
-function constructDiagnosticsPerFile(diagnostics: FormattedDisplayableDiagnostic[]): string {
-  const diagnosticsByFile = diagnostics.reduce<Record<string, string[]>>((acc, diag) => {
-    const file = diag.file || '';
-    if (!acc[file]) {
-      acc[file] = [];
-    }
-    acc[file].push(`[${diag.line}:${diag.col}] ${diag.message}`);
-    return acc;
-  }, {});
-
-  return Object.entries(diagnosticsByFile).reduce((acc, [fileName, diagnostics]) => {
-    return `${acc}${fileName}\n${diagnostics.join('\n')}\n\n`;
-  }, '');
-}
-
-const getConfig = async (): Promise<GraphQLSPConfig | undefined> => {
-  const cwd = process.cwd();
-  const tsconfigpath = path.resolve(cwd, 'tsconfig.json');
-
-  // TODO: Remove redundant read and move tsconfig.json handling to internal package
-  const root = (await resolveTypeScriptRootDir(tsconfigpath)) || cwd;
-
-  let tsconfigContents: string;
-  try {
-    const file = path.resolve(root, 'tsconfig.json');
-    tsconfigContents = await fs.readFile(file, 'utf-8');
-  } catch (error) {
-    console.error('Failed to read tsconfig.json in current working directory.', error);
-    return;
-  }
-
-  let tsConfig: TsConfigJson;
-  try {
-    tsConfig = parse(tsconfigContents) as TsConfigJson;
-  } catch (err) {
-    console.error(err);
-    return;
-  }
-
-  const config = getGraphQLSPConfig(tsConfig);
-  if (!config) {
-    console.error(`Could not find a "@0no-co/graphqlsp" plugin in your tsconfig.`);
-    return;
-  }
-
-  return config;
-};
 
 export default main;
