+++ packages/internal/package.json
@@ -53,9 +53,7 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@0no-co/graphql.web": "^1.0.5",
-    "@0no-co/graphqlsp": "^1.8.0",
-    "ts-morph": "^22.0.0"
+    "@0no-co/graphql.web": "^1.0.5"
   },
   "peerDependencies": {
     "typescript": "^5.0.0",
